Step3: 리뷰 반영
- 매직넘버 상수화
- RacingGame 의 attempt 함수 테스트 가능하도록 수정
- RacingCarTest -> CarTest, RacingGameTest 로 분리
