+++ src/main/kotlin/racingcar/domain/RacingGame.kt
@@ -1,7 +1,5 @@
 package racingcar.domain
 
-import kotlin.random.Random
-
 class RacingGame(
     val carSize: Int,
     val attemptSize: Int
@@ -13,13 +11,17 @@ class RacingGame(
     val history: List<List<Int>>
         get() = _history
 
-    fun attempt() {
+    fun attempt(conditions: List<Int>) {
         check(curAttemptNum < attemptSize) { "시도 횟수를 초과했습니다." }
 
-        cars.forEach {
-            it.move(Random.nextInt(10))
+        cars.forEachIndexed { index, car ->
+            car.move(conditions[index])
         }
         _history.add(cars.map { it.position })
         curAttemptNum++
+    }
+
+    companion object {
+        const val MAX_CONDITION: Int = 10
     }
 }
