+++ src/main/kotlin/racingcar/ui/main.kt
@@ -1,6 +1,7 @@
 package racingcar.ui
 
 import racingcar.domain.RacingGame
+import kotlin.random.Random
 
 fun main() {
     val inputView = InputView().also {
@@ -13,7 +14,11 @@ fun main() {
     val resultView = ResultView()
 
     repeat(inputView.attemptSize) {
-        racingGame.attempt()
+        racingGame.attempt(
+            List(racingGame.carSize) {
+                Random.nextInt(RacingGame.MAX_CONDITION)
+            }
+        )
     }
     resultView.printRacingGameHistory(racingGame.history)
 }
